Fixing my messed up shuttle PR (#11210)
# About the pull request
@Blundir 
Fixes #11151 

So I seem to have messed up in the testing phase and was on the wrong
branch. The chem-goons and bodyguards spawned on the old ERT base
instead of on a shuttle in the air on it's way to the Almayer. This PR
is a quick hotfix to this issue.

# Changelog
:cl:
fix: The Chem-Goons and Bodyguards now spawn properly.
/:cl: