Medical Physical Status QOL (#11183)
# About the pull request
Adds more medical physical status options to better reflect the physical
status of a marine in regards to their duties.

# Explain why it's good for the game
This will help doctors keep track of their patience better and in more
detail.

# Changelog

:cl:
qol: Medical Physical Status can now be changed to better reflect the
status of a marine.
/:cl: