Gives FTLs RTO Packs (#11185)
# About the pull request

FTLs now get RTO bags by default.

# Explain why it's good for the game

With the addition of #11024, secondary ways of getting important
information to and from CIC can get hairy - especially in lowpop. This
change should make FTL's an important facet of communication for the SL
to use when needed.

The Bag also only costs 5 points - so Balance-wise this shouldnt be too
large an issue.

# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl:
add: FTLs now get a free RTO Bag.
/:cl: