Fixes IMS not applying open chest/head overlay sprites (#11269)
# About the pull request

Fixes #11206 
IMS wasn't applying the overlay a retractor applies when widening the
incision on the chest/head, so I fixed it.

# Explain why it's good for the game

How are we supposed to know the chest and head are open if the overlay
doesn't show up? Missing overlays need fixing.
# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>


![dreamseeker_2025-12-04_19-16-13](https://github.com/user-attachments/assets/fc9b5b31-9991-4dfb-9e8d-dc061cb492a3)


</details>


# Changelog

:cl: Puckaboo2
fix: Using the IMS on the head or chest now applies their open incision
overlays.
/:cl: