adds hunted xeno hive, disables drone castes in hunting grounds (#11265)
# About the pull request

adds a hive used in hunting ground that does not allow drone evolutions

# Explain why it's good for the game

the hunting grounds are for fighting and all the castes in the drone
line are already not huntable on the normal map. all they really do is
be annoying and make xenos a pain in the ass to fight, leading to them
never being called.


# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl: Drathek, ori
add: added a new hive type used in the hunting grounds
balance: you can no longer evolve to drone or drone evolutions in the
hunting grounds
/:cl: