Facehuggers no longer eat your glasses or cigarettes (#11047)
# About the pull request

Title.
This is for pugie.

# Explain why it's good for the game

Facehuggers lacked a check for the mask slot to drop any other item that
isn't a mask to the ground, deleting it or putting it in null space
instead

This fixes that, now facehuggers no longer has a diet of ray bans and
cigarettes.

# Testing Photographs and Procedure
it works, you really don't need a video for this

# Changelog
:cl:
fix: Facehuggers no longer has a diet of ray bans and cigarettes, they
will no longer eat your glasses or cigarettes on your face slot
/:cl: