Rebalances Yautja mask knock off chance. (#11280)
# About the pull request

As title

# Explain why it's good for the game
Makes it more dynamic, and less likely for a crit pred to get
accidentally killed.

# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl:
balance: Yautja masks now fall off at different rates if human or in
crit.
fix: The 15% chance max for Yautja mask fall off is now functional.
/:cl:

Co-authored-by: forest2001 <forest2001@cm-ss13.com>