Adds 'soundtweak' and 'icontweak' changelog flags (#11309)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

Added two new changelog flags: 'soundtweak' for if an existing sound
file has been modified, and 'icontweak' for if an existing icon has been
modified.

# Explain why it's good for the game
We have a 'maptweak'. Why not have a 'soundtweak' and 'icontweak', as
well?

Most often, the sprites I change are simply tweaks. It feels the times I
actually DO make a new addition to the game, it may be looked over as
another tweak unless somebody checks the icon bot for added files.

Same goes for soundtweaks, in the future.

<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog

<!-- If your PR modifies aspects of the game that can be concretely
observed by players or admins you should add a changelog. If your change
does NOT meet this description, remove this section. Be sure to properly
label your changes in the changelog. Please note that maintainers freely
reserve the right to remove and add tags should they deem it
appropriate. You can attempt to finagle the system all you want, but
it's best to shoot for clear communication right off the bat. -->
<!-- If you add a name after the ':cl', that name will be used in the
changelog. You must add your CKEY after the CL if your GitHub name
doesn't match. Maintainers freely reserve the right to remove and add
tags should they deem it appropriate. -->

:cl: Puckaboo2
add: Adds 'soundtweak', and 'icontweak' in the changelog.
code: Alphabetizes changelog flags.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! -->

---------

Co-authored-by: Drathek <76988376+Drulikar@users.noreply.github.com>