Invert tgui alert yes/no options to be consistent (#11322)
# About the pull request
Title
Anything that had Yes and then No is now 
No and then Yes
Resolves: #11242
<!-- Remove this text and explain what the purpose of your PR is.

Mention if you have tested your changes. If you changed a map, make sure
you used the mapmerge tool.
If this is an Issue Correction, you can type "Fixes Issue #169420" to
link the PR to the corresponding Issue number #169420.

Remember: something that is self-evident to you might not be to others.
Explain your rationale fully, even if you feel it goes without saying.
-->

# Explain why it's good for the game
Organises the Left option to consistantly be Yes, because that's what
the user expects from 70%+ of these.
# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl: MistChristmas
qol: Give Obj Confirmation box now has Yes listed as the highlighted
Option on the right. A few other similar confirmation boxes got this
treatment. Including some Admin Verbs.
/:cl: