Immersive rank pins + Army Framework  (#11121)
# About the pull request
### Needs testing.

The pull request does as it's title claims, it adds sprites for
individual ranks pins upwards.

Additionally there are the full inspiration sprites, that are currently
unused but included in the code for future uses or for re-spriting
attempts as references.

O1 to O11(or O10S as it's refered to in code) share the same sprite
across branches, except for the navy which I will likely make shoulder
ranks and pads for in the future.

The army framework is simply the addition of a army faction in the pay
grades section and in the Rank-pins.dm. It's made to make adding sprites
sprites to the army branch and for easier merging of [a map that has
already made the army a
faction.](https://github.com/cmss13-devs/cmss13/pull/8846/commits)

### Needs testing.

# Explain why it's good for the game

It adds to the game's immersion without modifying too many things, and
being somewhat tame.

Plus it standardises the army faction stuff, so that it's easier to add
to it later on.

# Testing Photographs and Procedure

<details>
<summary>Screenshots & Videos</summary>

<img width="459" height="459" alt="Rankpins lineup"
src="https://github.com/user-attachments/assets/bd6ad119-a932-43c5-8adf-ea0a8a55ced7"
/>
<img width="1687" height="345" alt="rankpinsonmob"
src="https://github.com/user-attachments/assets/281eaad0-d02f-486f-913e-dc4d86d151f6"
/>


Ignore the round thing and square thing that are not actually used or
implemented, they were my first drafts at making the Non-descript broche
that E-1 uses in the army and in the marines. Since they don't have an
insignia.

</details>


# Changelog

Add: cmss13/code/datums/paygrades/factions/army
deleted: cmss13/code/datums/paygrades/factions/army due to conflict
issues
Add: cmss13/code/clothing/under/rank_pins.dm
Imageadd: cmss13/icons/obj/items/clothing/accessory/ranks.dmi

:cl:
add: added a few lines to Rank_pins.dm, mostly ones to make them use the
new sprites.
add: A army section to rank_pins.dm, to make it easier to change and add
stuff for that faction once it's implemented.
imageadd: Added a few icons to the Ranks.dmi
imageadd: Added a few icons to the Onmob ranks.dmi
imageadd: Added a few icons to the  Inventory ranks.dmi
/:cl:

<!-- Both :cl:'s are required for the changelog to work! -->